--- a/packages/@glimmer/runtime/lib/environment.ts
+++ b/packages/@glimmer/runtime/lib/environment.ts
@@ -137,7 +137,9 @@ export class EnvironmentImpl implements Environment {
       'A glimmer transaction was begun, but one already exists. You may have a nested transaction, possibly caused by an earlier runtime exception while rendering. Please check your console for the stack trace of any prior exceptions.'
     );
 
-    prepareTagsIfNeeded();
+    setTimeout(() => {
+      prepareTagsIfNeeded();
+    });
     this.debugRenderTree?.begin();
 
     this[TRANSACTION] = new TransactionImpl();