This PR is create to check GC behaviour on framework benchmark if we start using tags and transactions pools.
Where transaction pool is retainable and tags pool is predefined

Idea here is:
* trying to reuse `Tracker` classes, including it's `Sets` and prepare initial pool (reusable)
* Prepare `DirtyableTag` initial  pool (not reusble)
* Prepare `UpdatableTag` initial pool (not reusable)

In abstract ember app we always need some amount of `Tracker`, `DirtyableTag` and `UpdatableTag` classes.
And it may be a case where we will see GC improvements. And will be able to ship default pool sizes or allow developers to customize it from global variable or code replace functionality or babel plugin.