@lifeart 2023-12-22
locally, on 20 runs: (10k items rendering)

```
 Benchmark Results Summary    

duration phase estimated improvement -50ms [-64ms to -37ms] OR -1.66% [-2.12% to -1.25%]
renderStart phase estimated improvement -47ms [-61ms to -39ms] OR -2.7% [-3.47% to -2.2%]
glimmer-render-10000-rows-start phase no difference [-6ms to 7ms]
paint phase no difference [-1ms to 1ms]
```