@lifeart 2023-12-24
### Dec 25, updated benchmark

<img width="773" alt="image" src="https://github.com/glimmerjs/glimmer-vm/assets/1360552/39f83d0c-7316-41c7-b681-a8ae5e4860b0">

```
    Benchmark Results Summary    

duration phase estimated regression +345ms [246ms to 441ms] OR +2.09% [1.49% to 2.67%]
renderEnd phase estimated improvement -3ms [-4ms to -1ms] OR -3.77% [-6.16% to -1.45%]
render1000Items1End phase estimated improvement -35ms [-47ms to -25ms] OR -2.73% [-3.64% to -1.95%]
clearItems1End phase no difference [-2ms to 3ms]
render1000Items2End phase estimated improvement -11ms [-18ms to -1ms] OR -1.03% [-1.77% to -0.13%]
clearItems2End phase estimated improvement -2ms [-4ms to 0ms] OR -3.25% [-6.38% to -0.71%]
render5000Items1End phase no difference [-32ms to 38ms]
clearManyItems1End phase estimated improvement -3ms [-6ms to -1ms] OR -1.32% [-2.43% to -0.21%]
render5000Items2End phase estimated regression +430ms [251ms to 520ms] OR +10.25% [5.97% to 12.4%]
clearManyItems2End phase estimated improvement -7ms [-10ms to -4ms] OR -2.82% [-4.16% to -1.52%]
render1000Items3End phase estimated improvement -19ms [-34ms to -16ms] OR -2.18% [-3.94% to -1.83%]
append1000Items1End phase estimated improvement -12ms [-20ms to 0ms] OR -1.27% [-2.02% to -0.05%]
append1000Items2End phase estimated improvement -17ms [-26ms to -10ms] OR -1.73% [-2.61% to -1%]
updateEvery10thItem1End phase no difference [-6ms to 2ms]
updateEvery10thItem2End phase no difference [-1ms to 0ms]
selectFirstRow1End phase no difference [-1ms to 0ms]
selectSecondRow1End phase estimated regression +6ms [0ms to 10ms] OR +4.48% [0.18% to 7.02%]
removeFirstRow1End phase estimated regression +12ms [8ms to 17ms] OR +4.65% [3.15% to 6.67%]
removeSecondRow1End phase estimated regression +13ms [7ms to 16ms] OR +5.22% [2.94% to 6.39%]
swapRows1End phase no difference [-2ms to 1ms]
swapRows2End phase no difference [0ms to 8ms]
clearItems4End phase estimated regression +13ms [10ms to 16ms] OR +9.25% [7.25% to 11.03%]
paint phase no difference [0ms to 0ms]
```
[artifact-1.pdf](https://github.com/glimmerjs/glimmer-vm/files/13762861/artifact-1.pdf)